requirements/dev.txt
Flask
python-dotenv
# Add other development dependencies here